+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `ChromoteSession$screenshot()` gains an `options` argument that accepts a list of additional options to be passed to the Chrome Devtools Protocol's [`Page.captureScreenshot` method](https://chromedevtools.github.io/devtools-protocol/tot/Page/#method-captureScreenshot). (#129)
 
+* `ChromoteSession$screenshot()` will now infer the image format from the `filename` extension. Alternatively, you can specify the `format` in the list passed to `options`. (#130)
+
 # chromote 0.1.2
 
 * Fixed #109: An error would occur when a `Chromote` object's `$close()` method was called. (#110)